requirements.txt
numpy
torch
torchvision
timm
optuna
scikit-learn
pillow
seaborn